requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.0
pymongo==4.13.2
pandas==1.5.3
numpy==1.24.4
scikit-learn==1.7.0
requests==2.32.4
gdown==5.2.0

